chore: move type coercion functions to utils for better organization and reuse
